moved components into their own files, structured project
