+++ src/components/timeRow.js
@@ -1,0 +1,7 @@
+import React, { Component } from "react";
+
+export class TimeRow extends Component {
+  render() {
+    return <tr>{this.props.children}</tr>;
+  }
+}
